vou_viajar/requirements.txt

Django==2.1.3
djangorestframework==3.7.7
django-model-utils
Pillow==5.0.0
PyJWT==1.5.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
urllib3==1.22